fix(#42 WI-9a): Gate-4 audit fixes (2 Medium)
- Nav dispatch Tasks now re-read `boundNavigator` INSIDE the Task with
  `[weak self]` instead of capturing the navigator before the async hop — so a
  detach()/dismantle racing between the intent firing and the task executing
  makes the task no-op rather than driving a torn-down navigator.
- Split the 495-line ReadiumEPUBHost.swift into three files (host View 227 /
  ReadiumReaderCoordinator 176 / ReadiumNavigatorRepresentable 123 lines) — all
  under the ~300 convention, and cohesive as the Readium engine grows through
  WI-10+. Representable dropped `private` to move out; no behavior change.

104 tests / 7 suites green.
